style: apply forge fmt on src contracts

+++ contracts/interfaces/liquity/IHintHelpers.sol
@@ -3,26 +3,17 @@
 pragma solidity =0.8.24;
 
 interface IHintHelpers {
-
-    function getRedemptionHints(
-        uint _LUSDamount, 
-        uint _price,
-        uint _maxIterations
-    )
+    function getRedemptionHints(uint256 _LUSDamount, uint256 _price, uint256 _maxIterations)
         external
         view
-        returns (
-            address firstRedemptionHint,
-            uint partialRedemptionHintNICR,
-            uint truncatedLUSDamount
-        );
+        returns (address firstRedemptionHint, uint256 partialRedemptionHintNICR, uint256 truncatedLUSDamount);
 
-    function getApproxHint(uint _CR, uint _numTrials, uint _inputRandomSeed)
+    function getApproxHint(uint256 _CR, uint256 _numTrials, uint256 _inputRandomSeed)
         external
         view
-        returns (address hintAddress, uint diff, uint latestRandomSeed);
+        returns (address hintAddress, uint256 diff, uint256 latestRandomSeed);
 
-    function computeNominalCR(uint _coll, uint _debt) external pure returns (uint);
+    function computeNominalCR(uint256 _coll, uint256 _debt) external pure returns (uint256);
 
-    function computeCR(uint _coll, uint _debt, uint _price) external pure returns (uint);
+    function computeCR(uint256 _coll, uint256 _debt, uint256 _price) external pure returns (uint256);
 }
